[FLINK-6384] [py] Remove python binary check via additional process
The PythonStreamer used to check for the existence of the python binary by
starting a python process. This process was not closed afterwards. This caused
the PythonPlanBinderTest to fail locally.

I think the check whether a python binary exists is not necessary since the
subsequent python command would fail anyway if there is no binary available on
the system. The system failure message is that there is no such file or directory.
This error message should be descriptive enough in order to debug such a problem.

This closes #3774.