[FLINK-6792] [mvn] Move maven-checkstyle-plugin before maven-shade-plugin definition in pom.xml
Somehow appending the maven-checkstyle-plugin at the end of the pom.xml changed the order
of the maven-assembly-plugin and the maven-shade-plugin in the package phase. This breaks
things since the shade plugin has to run before the assembly plugin to build the fat-jar.
Specifying the checkstyle-plugin before the maven-shade-plugin resolves the issue.

This closes #4042.