[FLINK-7011] [kafka] Remove Kafka testStartFromKafkaCommitOffsets ITCases
The testStartFromKafkaCommitOffsets ITCases are covering too much within
one single test. The case verifies that whatever offset was committed to
Kafka, Flink rreads it correctly and can uses that as the correct
starting point for exactly-once.

This over-engineered test was instable that we needed to first read some
records and wait until some records is committed. This wait is hard to
define.

It is in fact sufficient to have 2 separate tests to cover the tested
behaviour:
- test that committed Kafka offsets are correct (there is already a
  ITCase for this, i.e. `runCommitOffsetsToKafka`)
- test that committed offsets are correctly picked up and used correctly
  (there is actually also a test for this, i.e.
`runStartFromGroupOffsets`)

Hence, this test can be removed without harming test coverage.

This closes #4190.