[FLINK-7354][tests] ignore "initialSeedUniquifierGenerator" thread in thread list
Netty may spawn a thread in its ThreadLocalRandom class because of some secure
random use which may be caught in our LocalFlinkMiniClusterITCase as not being
stopped. Let's tolerate this.

Alternatively, we could solve this similarly to
https://issues.apache.org/jira/browse/SOLR-10098 by setting
ThreadLocalRandom.setInitialSeedUniquifier(1L);
but that may be less future proof or even remove some randomness.

This closes #4464.