[FLINK-7240] [tests] Stabilize ExternalizedCheckpointITCase
The problem was that the TestingCluster did not wait properly after canceling the
job that the job was also completely removed from the cluster before submitting
the next job. This could lead to a NoResourceAvailableException which ultimately
made the job fail.

Add CountDownLatch to source of ExternalizedCheckpointITCase in order to wait for all sources to be running before triggering a checkpoint

This closes #4497.