[FLINK-7500] Set JobMaster leader session id in main thread
This commit changes the JobMaster such that it's leader session id is only set
in the main thread. This is done by passing the leader session id to the
startJobExecution method. Moreover, this commit returns a future from the
start and suspend methods which can be used to wait on the completion of these
operations.

This closes #4579.