[FLINK-2268] Don't include Hadoop deps in flink-core/flink-java
This also makes them optional in flink-runtime, which is enabled by the
previous changes to only use Hadoop dependencies if they are available.

This also requires adding a few explicit dependencies in other modules
because they were using transitive dependencies of the Hadoop deps. The
most common dependency there is, ha!, commons-io.