[FLINK-7556] Allow Integer.MIN_VALUE for fetch size in JDBCInputFormat
Allow Integer.MIN_VALUE to be accepted as a parameter for setFetchSize for MySQL Driver.

The combination of a forward-only, read-only result set, with a fetch size of Integer.MIN_VALUE serves as a signal to the driver to stream result sets row-by-row. After this, any result sets created with the statement will be retrieved row-by-row.

This closes #4617.