[FLINK-7072] [REST] Define protocol for job submit/cancel/stop
[FLINK-7072] [REST] Extend Dispatcher

[FLINK-7072] [REST] Add handlers for job submit/cancel/stop

[FLINK-7072] [REST] CLI integration

use ExecutorThradFactory + rebase(blobKey fix)

add "Flink" prefix to RestCC threads

shutdown client for cancel/shutdown

Rework CliFrontEnd Stop/Cancel tests

These tests verified that the CLI was sending the correct messages and
parameters to the JM actor. This is now handled by the ClusterClient, so
the tests were adjusted to verify that the correct methods on the
ClusterClient are being called.

Additional tests were added to the ClusterClientTest class to verify
that the correct messages and parameters are being sent.

This closes #4742.