[FLINK-5706] [file systems] Add S3 file systems without Hadoop dependencies
This adds two implementations of a file system that write to S3.
Both are not actual re-implementations but wrap other implementations and shade dependencies.

(1) A wrapper around Hadoop's s3a file system. By pulling a smaller dependency tree and
    shading all dependencies away, this keeps the appearance of Flink being Hadoop-free,
    from a dependency perspective.

(2) The second S3 file system is from the Presto Project.
    Initial simple tests seem to indicate that it responds slightly faster
    and in a bit more lightweight manner to write/read/list requests, compared
    to the Hadoop s3a FS, but it has some semantical differences.

This closes #4818