[hotfix] [scala api] Move tests to correct package
We previously had all Scala API unit tests in the flink-tests
project, because Eclipse could not use macros in 'test' that were
declared in 'main'.

Because we do not support Eclipse for development of the system any
more (only for using Flink to develop Flink-based applications),
we can now move the tests to their natural location and simplify
some of the dependency structures.