[build] Move licenses for shaded dependencies into the NOTICE file
In accordance with the Apache Software License 2.0, the NOTICE
file is the dedicated/prefered place for any additional license
information that affects redistribution, because the NOTICE
file is required to be included in any redistribution.

The only project still having shaded dependencies that require
extra license statements is flink-table. Changing the licenses
to be in the NOTICE file also allows removing the maven-resource
plugin from the root pom.xml