[FLINK-7637] [kinesis] Fix at-least-once guarantee in FlinkKinesisProducer
Prior to this commit, there is no flushing of KPL outstanding records on
checkpoints in the FlinkKinesisProducer. Likewise to the at-least-once
issue on the Flink Kafka producer before, this may lead to data loss if
there are asynchronous failing records after a checkpoint which the
records was part of was completed.

This closes #4871.