[FLINK-7924] [checkpoints] Fix incorrect names of checkpoint options
Checkpoint options are incorrectly always called 'FULL_CHECKPOINT' when actually,
the checkpoints may always be incremental and only savepoints have to be full
and self contained.

Initially, we planned to add options for multiple checkpoints, like checkpoints
that were foreced to be full, and checkpoints that were incremental. That
is not necessary at this point.