[hotfix] [kinesis] Use ArrayList for faster shard state updates
Previously, the Kinesis Consumer uses LinkedLists as the underlying
implementation for subscribedShardsState. This list is accessed on every
record, updating a shard's state using a shard state index (i.e., the
order of the shard state in the list). With LinkedLists, this access has
linear time complexity, and since this operation happens per record, can
have a big performance hit on executions with a huge number of Kinesis
shards.

This commit changes the list implementation to be an ArrayList for
constant-time shard state access. The downside is that when new shards
are discovered, there will be a costly array re-size. However, since
resharding is not expected to happen so often, this is acceptable.