[STRMCMP-860] Fix thread cancelation timeout problems
This attempts to fix the hanging task thread. Two possible scenarios:

(1) `cancelTask()` can be called from multiple locations. At least one of the
locations does not perform proper exception handling. Once a task is in the
CANCELING state and the cancelation timeout thread has not been started, the
task will linger forever.

(2) `taskCancellationTimeout` was non-volatile and is non-final which
theoretically could lead to a thread calling `cancelTask()` to read an outdated
value. It is first initialized with 0 which would disable the cancelation. The
value is written in the constructor which makes it unlikely but not impossible
to read the initial value:
https://stackoverflow.com/questions/33485625/non-volatile-fields-first-object-access-from-another-thread-java

We bring up the thread as soon as possible to avoid any problems running
it. Also we exit in case the thread fails to start.