[STRMCMP-860][FLINK-14949][FLINK-16573][flink 1.9.3, 1.10.1] Fix thread cancelation timeout problems (#39)
* [STRMCMP-860] Fix thread cancelation timeout problems

This attempts to fix the hanging task thread. Two possible scenarios:

(1) `cancelTask()` can be called from multiple locations. At least one of the
locations does not perform proper exception handling. Once a task is in the
CANCELING state and the cancelation timeout thread has not been started, the
task will linger forever.

(2) `taskCancellationTimeout` was non-volatile and is non-final which
theoretically could lead to a thread calling `cancelTask()` to read an outdated
value. It is first initialized with 0 which would disable the cancelation. The
value is written in the constructor which makes it unlikely but not impossible
to read the initial value:
https://stackoverflow.com/questions/33485625/non-volatile-fields-first-object-access-from-another-thread-java

We bring up the thread as soon as possible to avoid any problems running
it. Also we exit in case the thread fails to start.

* Make sure to always shutdown forcefully

* Catch all errors during task cancellation

* [FLINK-16573] Ensure Kinesis RecordFetcher threads shutdown on cancel

The threads may not shut down correctly because they do not check for the
running flag in the inner loops. The threads also do not get interrupted because
they are not connected to the main task thread.

These threads keep lingering around after the job has shut down:

```
Thread 23168: (state = BLOCKED)
 - java.lang.Object.wait(long) @bci=0 (Compiled frame; information may be imprecise)
 - org.apache.flink.streaming.connectors.kinesis.util.RecordEmitter.emitRecords() @bci=140, line=209 (Compiled frame)
 - org.apache.flink.streaming.connectors.kinesis.util.RecordEmitter.run() @bci=18, line=177 (Interpreted frame)
 - java.lang.Thread.run() @bci=11, line=748 (Compiled frame)
```